Change e_chil.c to use ltdl instead of OpenSSL's DSO functions
From OpenSSL 1.1.0, OpenSSL's DSO functions are no longer public, so
we need to use something else.  Considering all the variants that
exist, it makes sense to use a convenience layer instead of inventing
our own, and since we already use libtool, we might as well use
libltdl, which comes with it.